requirements.txt
python-dotenv

streamlit
numpy
scipy

matplotlib
scikit-learn

tqdm

langchain-openai

fasttext
gensim

transformers
torch